Email: Move `EmailType` from `enums` to `integrations/email/typing`
