+++ backend/tests/test_integrations/test_email/test_email_types/test_permission_deactivation.py
@@ -11,10 +11,11 @@ from django.test import override_settings
 from freezegun import freeze_time
 
 from tilavarauspalvelu.admin.email_template.utils import get_mock_data
-from tilavarauspalvelu.enums import EmailType, Language
+from tilavarauspalvelu.enums import Language
 from tilavarauspalvelu.integrations.email.main import EmailService
 from tilavarauspalvelu.integrations.email.rendering import render_html, render_text
 from tilavarauspalvelu.integrations.email.template_context import get_context_for_permission_deactivation
+from tilavarauspalvelu.integrations.email.typing import EmailType
 from utils.date_utils import local_datetime
 
 from tests.factories import UnitFactory, UserFactory
