+++ backend/tests/test_integrations/test_email/test_email_types/test_seasonal_reservation_modified_single.py
@@ -11,10 +11,11 @@ from django.test import override_settings
 from freezegun import freeze_time
 
 from tilavarauspalvelu.admin.email_template.utils import get_mock_data, get_mock_params
-from tilavarauspalvelu.enums import EmailType, ReservationStateChoice, ReservationTypeChoice
+from tilavarauspalvelu.enums import ReservationStateChoice, ReservationTypeChoice
 from tilavarauspalvelu.integrations.email.main import EmailService
 from tilavarauspalvelu.integrations.email.rendering import render_html, render_text
 from tilavarauspalvelu.integrations.email.template_context import get_context_for_seasonal_reservation_modified_single
+from tilavarauspalvelu.integrations.email.typing import EmailType
 
 from tests.factories import ReservationFactory
 from tests.helpers import TranslationsFromPOFiles
